requirements.txt
pytest
aws-cdk-lib
constructs
black
pre-commit
boto3
pydantic
aws_cdk.aws_lambda_python_alpha
aws_lambda_powertools
